+++ web/tests/e2e/onboarding_flow.spec.ts
@@ -2,19 +2,47 @@ import { expect, test } from "@playwright/test";
 import type { Page } from "@playwright/test";
 import { loginAs, loginAsRandomUser, apiLogin } from "./utils/auth";
 import { OnyxApiClient } from "./utils/onyxApiClient";
-import { expectElementScreenshot } from "./utils/visualRegression";
+import {
+  expectScreenshot,
+  expectElementScreenshot,
+} from "./utils/visualRegression";
 
 /**
- * Onboarding Flow E2E Tests
+ * Onboarding E2E Tests
  *
- * Tests the 4 main user scenarios:
- * 1. Admin WITHOUT LLM providers -> Full onboarding, chat disabled
- * 2. Admin WITH LLM providers -> No full onboarding, chat enabled
- * 3. Non-admin WITHOUT LLM providers -> NonAdminStep name prompt, chat disabled
- * 4. Non-admin WITH LLM providers -> NonAdminStep name prompt, chat enabled
+ * Covers the four main user scenarios:
+ *   1. Admin  WITHOUT LLM providers → Full 4-step onboarding, chat disabled
+ *   2. Admin  WITH    LLM providers → No full onboarding, simple name prompt
+ *   3. Non-admin WITHOUT LLM providers → Name prompt only, chat disabled
+ *   4. Non-admin WITH    LLM providers → Name prompt only, chat enabled
  *
- * Marked @exclusive because scenarios 1 & 3 delete all LLM providers.
+ * The admin onboarding flow (Welcome → Name → LLM Setup → Complete) appears
+ * on `/app` for admin users who:
+ *   1. Have zero chat sessions, AND
+ *   2. Have no LLM providers configured (`hasAnyProvider === false`).
+ *
+ * Both conditions are enforced by `useShowOnboarding` in
+ * `web/src/hooks/useShowOnboarding.ts`.  There is no localStorage key
+ * involved — the check is purely server-state-driven.
+ *
+ * Non-admin users see a simpler single-step name prompt whenever
+ * `user.personalization.name` is not set, regardless of LLM providers.
+ *
+ * Marked @exclusive because scenarios 1 & 3 delete all LLM providers,
+ * which is shared backend state.  The exclusive Playwright project runs
+ * tests serially in a single worker to prevent provider-mutation races.
+ *
+ * NOTE: Many text elements on the onboarding page are rendered via the
+ * `Truncated` component, which places a hidden offscreen copy for width
+ * measurement.  `getByText(...)` can therefore resolve to 2 elements —
+ * we use `.first()` where needed to avoid strict-mode violations.
  */
+
+const THEMES = ["light", "dark"] as const;
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
 
 async function deleteAllProviders(client: OnyxApiClient): Promise<void> {
   const providers = await client.listLlmProviders();
@@ -32,11 +60,10 @@ async function deleteAllProviders(client: OnyxApiClient): Promise<void> {
 async function createFreshAdmin(
   page: Page
 ): Promise<{ email: string; password: string }> {
-  // First, log in as the existing admin so we can promote the new user
   await page.context().clearCookies();
   const { email, password } = await loginAsRandomUser(page);
 
-  // Now promote the new user to admin via the existing admin
+  // Promote the new user to admin via the pre-provisioned admin
   await page.context().clearCookies();
   await loginAs(page, "admin");
   const adminClient = new OnyxApiClient(page.request);
@@ -56,26 +83,120 @@ async function createFreshUser(
   return await loginAsRandomUser(page);
 }
 
-test.describe("Onboarding Flow @exclusive", () => {
-  test.describe("Scenario 1: Admin WITHOUT LLM providers", () => {
+/**
+ * Connect an LLM provider through the onboarding form UI.
+ *
+ * The "Next" button on the LLM step is disabled until a provider is connected
+ * via the form (which calls `setButtonActive(true)` internally).  We cannot
+ * simply create a provider via the API because `useLLMProviders` has
+ * `dedupingInterval: 60 000` — SWR will not re-fetch for a full minute, so
+ * the button stays disabled.
+ *
+ * To avoid depending on an external LLM service being reachable, the
+ * `/api/admin/llm/test` endpoint is intercepted and returns 200 immediately.
+ */
+async function connectProviderViaForm(page: Page): Promise<void> {
+  // Mock the LLM key-test endpoint so the form succeeds without hitting OpenAI
+  await page.route("**/api/admin/llm/test", (route) =>
+    route.fulfill({
+      status: 200,
+      contentType: "application/json",
+      body: "{}",
+    })
+  );
+
+  // Wait for provider cards to finish loading
+  await expect(page.getByText("Connect", { exact: true }).first()).toBeVisible({
+    timeout: 10000,
+  });
+
+  // Click the first provider card (GPT / OpenAI) to open its connection form
+  await page
+    .locator('[role="button"]')
+    .filter({ hasText: "GPT" })
+    .first()
+    .click();
+
+  // Fill the API key — the test endpoint is mocked so any value works
+  const dialog = page.locator('[role="dialog"]');
+  await expect(dialog).toBeVisible({ timeout: 5000 });
+  await dialog.locator("input").first().fill("e2e-test-api-key");
+
+  // Click the modal's "Connect" submit button
+  await dialog.getByRole("button", { name: "Connect" }).click();
+
+  // Wait for the modal to close (provider created, button enabled)
+  await expect(dialog).not.toBeVisible({ timeout: 15000 });
+
+  // Clean up the route mock
+  await page.unroute("**/api/admin/llm/test");
+}
+
+/**
+ * Navigate from Welcome through to the Complete step:
+ *   Welcome → Name (fill "Ada Lovelace") → LLM Setup → connect provider → Complete
+ *
+ * Uses {@link connectProviderViaForm} to advance past the LLM step.
+ */
+async function navigateToCompleteStep(page: Page): Promise<void> {
+  // Welcome → Let's Go
+  await page
+    .getByRole("button", { name: "Let's Go" })
+    .click({ timeout: 15000 });
+
+  // Name step — fill and advance
+  const nameInput = page.getByPlaceholder("Your name");
+  await expect(nameInput).toBeVisible({ timeout: 10000 });
+  await nameInput.fill("Ada Lovelace");
+  const nextBtn = page.getByRole("button", { name: "Next" });
+  await expect(nextBtn).toBeEnabled({ timeout: 5000 });
+  await nextBtn.click();
+
+  // Wait for LLM step to render
+  await expect(page.getByText("Connect your LLM models").first()).toBeVisible({
+    timeout: 10000,
+  });
+
+  // Connect a provider through the onboarding form
+  await connectProviderViaForm(page);
+
+  // The form called setButtonActive(true), so Next is enabled
+  await expect(nextBtn).toBeEnabled({ timeout: 5000 });
+  await nextBtn.click();
+
+  // Wait for Complete step
+  await expect(page.getByText(/You're all set/).first()).toBeVisible({
+    timeout: 15000,
+  });
+}
+
+// ---------------------------------------------------------------------------
+// Test suite
+// ---------------------------------------------------------------------------
+
+test.describe("Onboarding E2E @exclusive", () => {
+  // ─── Scenario 1: Admin WITHOUT LLM providers ────────────────────────
+  test.describe("Admin without LLM providers", () => {
     test.beforeEach(async ({ page }) => {
-      // Delete all providers first (as existing admin)
+      // Delete all providers (as existing admin)
       await page.context().clearCookies();
       await loginAs(page, "admin");
       const adminClient = new OnyxApiClient(page.request);
       await deleteAllProviders(adminClient);
 
-      // Create a fresh admin user (no chat history)
+      // Create a fresh admin user (no chat history, no name)
       await createFreshAdmin(page);
     });
 
     test.afterEach(async ({ page }) => {
-      // Restore providers
+      // Restore at least one public provider for other test suites
       await page.context().clearCookies();
       await loginAs(page, "admin");
       const adminClient = new OnyxApiClient(page.request);
       await adminClient.ensurePublicProvider();
     });
+
+    // ── Functional tests ────────────────────────────────────────────
 
     test("shows full onboarding flow with Welcome step", async ({ page }) => {
       await page.goto("/app");
@@ -138,9 +259,284 @@ test.describe("Onboarding Flow @exclusive", () => {
         name: "onboarding-llm-step",
       });
     });
+
+    test("can complete full onboarding via provider form", async ({ page }) => {
+      await page.goto("/app");
+      await page.waitForLoadState("networkidle");
+
+      await navigateToCompleteStep(page);
+
+      await expect(page.getByText("Step 3 of 3").first()).toBeVisible();
+
+      const finishBtn = page.getByRole("button", { name: "Finish Setup" });
+      await expect(finishBtn).toBeVisible();
+      await finishBtn.click();
+
+      await expect(
+        page.getByText("Let's take a moment to get you set up.").first()
+      ).not.toBeVisible({ timeout: 5000 });
+
+      const chatInput = page.locator("#onyx-chat-input-textarea");
+      await expect(chatInput).toBeVisible({ timeout: 10000 });
+    });
+
+    test("Enter in name field advances to LLM step", async ({ page }) => {
+      await page.goto("/app");
+      await page.waitForLoadState("networkidle");
+
+      await page
+        .getByRole("button", { name: "Let's Go" })
+        .click({ timeout: 15000 });
+
+      const nameInput = page.getByPlaceholder("Your name");
+      await expect(nameInput).toBeVisible({ timeout: 10000 });
+      await nameInput.fill("Ada Lovelace");
+
+      const nextBtn = page.getByRole("button", { name: "Next" });
+      await expect(nextBtn).toBeEnabled({ timeout: 5000 });
+      await nameInput.press("Enter");
+
+      await expect(
+        page.getByText("Connect your LLM models").first()
+      ).toBeVisible({ timeout: 10000 });
+    });
+
+    test("onboarding does not reappear after Finish Setup", async ({
+      page,
+    }) => {
+      await page.goto("/app");
+      await page.waitForLoadState("networkidle");
+
+      await navigateToCompleteStep(page);
+
+      await page
+        .getByRole("button", { name: "Finish Setup" })
+        .click({ timeout: 10000 });
+
+      await expect(
+        page.getByText("Let's take a moment to get you set up.").first()
+      ).not.toBeVisible({ timeout: 10000 });
+
+      // Reload and verify onboarding stays dismissed
+      await page.reload();
+      await page.waitForLoadState("networkidle");
+
+      await expect(
+        page.getByText("Let's take a moment to get you set up.").first()
+      ).not.toBeVisible({ timeout: 5000 });
+
+      await expectScreenshot(page, {
+        name: "onboarding-light-persistence-check",
+        hide: ['[data-testid="onyx-logo"]'],
+      });
+    });
+
+    // ── Theme-variant visual regression tests ───────────────────────
+
+    for (const theme of THEMES) {
+      test.describe(`Visual regression — ${theme}`, () => {
+        test("Welcome step", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          const welcomeText = page
+            .getByText("Let's take a moment to get you set up.")
+            .first();
+          await expect(welcomeText).toBeVisible({ timeout: 15000 });
+          await expect(
+            page.getByRole("button", { name: "Let's Go" })
+          ).toBeVisible();
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-welcome-step`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("Name step empty", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          await page
+            .getByRole("button", { name: "Let's Go" })
+            .click({ timeout: 15000 });
+
+          const nameInput = page.getByPlaceholder("Your name");
+          await expect(nameInput).toBeVisible({ timeout: 10000 });
+          await expect(
+            page.getByText("What should Onyx call you?").first()
+          ).toBeVisible();
+          await expect(page.getByText("Step 1 of 3").first()).toBeVisible();
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-name-step-empty`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("Name step filled", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          await page
+            .getByRole("button", { name: "Let's Go" })
+            .click({ timeout: 15000 });
+
+          const nameInput = page.getByPlaceholder("Your name");
+          await expect(nameInput).toBeVisible({ timeout: 10000 });
+          await nameInput.fill("Ada Lovelace");
+
+          const nextBtn = page.getByRole("button", { name: "Next" });
+          await expect(nextBtn).toBeEnabled({ timeout: 5000 });
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-name-step-filled`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("LLM setup step", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          await page
+            .getByRole("button", { name: "Let's Go" })
+            .click({ timeout: 15000 });
+
+          const nameInput = page.getByPlaceholder("Your name");
+          await expect(nameInput).toBeVisible({ timeout: 10000 });
+          await nameInput.fill("Ada Lovelace");
+
+          const nextBtn = page.getByRole("button", { name: "Next" });
+          await expect(nextBtn).toBeEnabled({ timeout: 5000 });
+          await nextBtn.click();
+
+          await expect(
+            page.getByText("Connect your LLM models").first()
+          ).toBeVisible({ timeout: 10000 });
+          await expect(page.getByText("Step 2 of 3").first()).toBeVisible();
+
+          const connectLabels = page.getByText("Connect", { exact: true });
+          await expect(connectLabels.first()).toBeVisible({ timeout: 10000 });
+          await expect(
+            page.getByRole("link", { name: "View in Admin Panel" })
+          ).toBeVisible();
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-llm-setup-step`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("Complete step", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          await navigateToCompleteStep(page);
+
+          await expect(page.getByText("Step 3 of 3").first()).toBeVisible();
+          await expect(
+            page.getByText("Select web search provider").first()
+          ).toBeVisible();
+          await expect(
+            page.getByText("Enable image generation").first()
+          ).toBeVisible();
+          await expect(
+            page.getByText("Invite your team").first()
+          ).toBeVisible();
+
+          const finishBtn = page.getByRole("button", {
+            name: "Finish Setup",
+          });
+          await expect(finishBtn).toBeVisible();
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-complete-step`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("Finish Setup dismisses onboarding", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          await navigateToCompleteStep(page);
+
+          const finishBtn = page.getByRole("button", {
+            name: "Finish Setup",
+          });
+          await expect(finishBtn).toBeVisible({ timeout: 10000 });
+          await finishBtn.click();
+
+          await expect(
+            page.getByText("Let's take a moment to get you set up.").first()
+          ).not.toBeVisible({ timeout: 5000 });
+
+          const chatInput = page.locator("#onyx-chat-input-textarea");
+          await expect(chatInput).toBeVisible({ timeout: 10000 });
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-post-onboarding`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("Full page screenshots", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          await expect(
+            page.getByText("Let's take a moment to get you set up.").first()
+          ).toBeVisible({ timeout: 15000 });
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-full-page-welcome`,
+            fullPage: true,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+
+          await page.getByRole("button", { name: "Let's Go" }).click();
+          const nameInput = page.getByPlaceholder("Your name");
+          await expect(nameInput).toBeVisible({ timeout: 10000 });
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-full-page-name`,
+            fullPage: true,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+      });
+    }
   });
 
-  test.describe("Scenario 2: Admin WITH LLM providers", () => {
+  // ─── Scenario 2: Admin WITH LLM providers ───────────────────────────
+  test.describe("Admin with LLM providers", () => {
     test.beforeEach(async ({ page }) => {
       // Ensure provider exists
       await page.context().clearCookies();
@@ -161,9 +557,7 @@ test.describe("Onboarding Flow @exclusive", () => {
       ).not.toBeVisible({ timeout: 5000 });
     });
 
-    test("shows name prompt (NonAdminStep) when name not set", async ({
-      page,
-    }) => {
+    test("shows name prompt when name not set", async ({ page }) => {
       await page.goto("/app");
       await page.waitForLoadState("networkidle");
 
@@ -191,7 +585,8 @@ test.describe("Onboarding Flow @exclusive", () => {
     });
   });
 
-  test.describe("Scenario 3: Non-admin WITHOUT LLM providers", () => {
+  // ─── Scenario 3: Non-admin WITHOUT LLM providers ────────────────────
+  test.describe("Non-admin without LLM providers", () => {
     test.beforeEach(async ({ page }) => {
       // Delete all providers (as existing admin)
       await page.context().clearCookies();
@@ -260,7 +655,8 @@ test.describe("Onboarding Flow @exclusive", () => {
     });
   });
 
-  test.describe("Scenario 4: Non-admin WITH LLM providers", () => {
+  // ─── Scenario 4: Non-admin WITH LLM providers ───────────────────────
+  test.describe("Non-admin with LLM providers", () => {
     test.beforeEach(async ({ page }) => {
       // Ensure provider exists
       await page.context().clearCookies();
@@ -309,5 +705,104 @@ test.describe("Onboarding Flow @exclusive", () => {
       await expect(namePrompt).not.toBeVisible({ timeout: 5000 });
       await expect(confirmation).not.toBeVisible();
     });
+
+    // ── Theme-variant visual regression tests ───────────────────────
+
+    for (const theme of THEMES) {
+      test.describe(`Visual regression — ${theme}`, () => {
+        test("Non-admin name prompt", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          await expect(
+            page.getByText("What should Onyx call you?").first()
+          ).toBeVisible({ timeout: 15000 });
+
+          const nameInput = page.getByPlaceholder("Your name");
+          await expect(nameInput).toBeVisible();
+
+          const saveBtn = page.getByRole("button", { name: "Save" });
+          await expect(saveBtn).toBeVisible();
+          await expect(saveBtn).toBeDisabled();
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-nonadmin-initial`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("Non-admin name filled and saved", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          const nameInput = page.getByPlaceholder("Your name");
+          await expect(nameInput).toBeVisible({ timeout: 15000 });
+          await nameInput.fill("Jane Doe");
+
+          const saveBtn = page.getByRole("button", { name: "Save" });
+          await expect(saveBtn).toBeEnabled();
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-nonadmin-name-filled`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+
+          const responsePromise = page.waitForResponse(
+            (resp) =>
+              resp.url().includes("/api/user/personalization") &&
+              resp.request().method() === "PATCH" &&
+              resp.status() === 200
+          );
+          await saveBtn.click();
+          await responsePromise;
+
+          await expect(
+            page.getByText("What should Onyx call you?").first()
+          ).not.toBeVisible({ timeout: 10000 });
+
+          const chatInput = page.locator("#onyx-chat-input-textarea");
+          await expect(chatInput).toBeVisible({ timeout: 10000 });
+
+          await expectScreenshot(page, {
+            name: `onboarding-${theme}-nonadmin-saved`,
+            hide: ['[data-testid="onyx-logo"]'],
+          });
+        });
+
+        test("Non-admin Enter to save", async ({ page }) => {
+          await page.addInitScript(
+            (t: string) => localStorage.setItem("theme", t),
+            theme
+          );
+          await page.goto("/app");
+          await page.waitForLoadState("networkidle");
+
+          const nameInput = page.getByPlaceholder("Your name");
+          await expect(nameInput).toBeVisible({ timeout: 15000 });
+          await nameInput.fill("John Doe");
+
+          const responsePromise = page.waitForResponse(
+            (resp) =>
+              resp.url().includes("/api/user/personalization") &&
+              resp.request().method() === "PATCH" &&
+              resp.status() === 200
+          );
+          await nameInput.press("Enter");
+          await responsePromise;
+
+          await expect(
+            page.getByText("What should Onyx call you?").first()
+          ).not.toBeVisible({ timeout: 10000 });
+        });
+      });
+    }
   });
 });
